% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{mhg}
\alias{mhg}
\alias{mhg-package}
\title{Non-parametric rank enrichment test for binary data.}
\description{
This package implements a non-parametric enrichment test for
ranked binary lists. This is a general package suitable for many
applications. For example, it can be used for gene-set enrichment tests to
find enrichment for gene annotations.
}
\details{
Run the test with \code{\link{mhg_test}}

Plot the results with \code{\link{plot_mhg}}
}
\references{
Eden, E., Lipson, D., Yogev, S. & Yakhini, Z. Discovering motifs in ranked
lists of DNA sequences. PLoS Comput. Biol. 3, e39 (2007).
\url{http://dx.doi.org/10.1371/journal.pcbi.0030039}

Wagner, F. GO-PCA: An Unsupervised Method to Explore Biological
Heterogeneity Based on Gene Expression and Prior Knowledge. bioRxiv (2015).
\url{http://dx.doi.org/10.1101/018705}
}

